feat: change cards shown based on new container widths
