+++ components/markets/MarketScroll.tsx
@@ -1,22 +1,11 @@
-import { motion } from "framer-motion";
-import { useStore } from "lib/stores/Store";
 import { observer } from "mobx-react";
-import Link from "next/link";
 import { useEffect, useRef, useState } from "react";
-import { ChevronLeft, ChevronRight } from "react-feather";
 import { useResizeDetector } from "react-resize-detector";
-import MarketCard, { IndexedMarketCardData } from "./market-card";
+import MarketCard, { IndexedMarketCardData } from "./market-card/index";
+import HorizontalScroll from "components/ui/HorizontalScroll";
 
 const MarketScroll = observer(
-  ({
-    title,
-    markets,
-    showMarketsLink = true,
-  }: {
-    title: string;
-    markets: IndexedMarketCardData[];
-    showMarketsLink?: boolean;
-  }) => {
+  ({ title, markets }: { title: string; markets: IndexedMarketCardData[] }) => {
     const scrollRef = useRef<HTMLDivElement>();
     const [scrollLeft, setScrollLeft] = useState(0);
     const [scrollDirection, setScrollDirection] = useState<"left" | "right">(
@@ -27,7 +16,13 @@ const MarketScroll = observer(
     const gap = 30;
     const scrollMin = 0;
     const scrollMax = cardWidth * markets.length + gap * (markets.length - 1);
-    const cardsShown = Math.floor(containerWidth / (gap + cardWidth));
+    // const cardsShown = Math.floor(containerWidth / (gap + cardWidth));
+    const cardsShown =
+      containerWidth < 716
+        ? 3
+        : containerWidth >= 716 && containerWidth < 1183
+        ? 2
+        : 3;
     const moveSize = cardsShown * (cardWidth + gap);
 
     useEffect(() => {
@@ -59,60 +54,40 @@ const MarketScroll = observer(
       hasReachedEnd || cardWidth * markets.length < containerWidth;
 
     return (
-      <div ref={containerRef} className="flex flex-col">
-        <div className="flex items-center mb-ztg-30">
-          <div className=" font-bold text-[28px]">{title}</div>
-          <div className="flex ml-auto items-center">
-            {showMarketsLink && (
-              <Link
-                href="markets"
-                className="text-ztg-14-150 border-2 border-pastel-blue rounded-[5px] px-[10px] py-[3px]"
-              >
-                Go To Markets
-              </Link>
-            )}
-            <button
-              onClick={handleLeftClick}
-              className={`hidden sm:flex items-center justify-center w-[26px] h-[26px] rounded-full ml-[12px] mr-[8px] ztg-transition ${
-                leftDisabled
-                  ? "bg-geyser text-pastel-blue"
-                  : "bg-pastel-blue text-white"
-              }`}
-              disabled={leftDisabled}
-            >
-              <ChevronLeft className="relative right-[1px]" />
-            </button>
-            <button
-              onClick={handleRightClick}
-              className={`hidden sm:flex items-center justify-center w-[26px] h-[26px] rounded-full ztg-transition  ${
-                rightDisabled
-                  ? "bg-geyser text-pastel-blue"
-                  : "bg-pastel-blue text-white"
-              }`}
-              disabled={rightDisabled}
-            >
-              <ChevronRight className="relative left-[1px]" />
-            </button>
-          </div>
-        </div>
-        <div className="relative">
+      <div ref={containerRef} className="grid sm:grid-cols-2 gap-4 md:gap-6">
+        <h3 className="sm:col-span-1 font-bold text-[28px]">{title}</h3>
+        <HorizontalScroll
+          classes="order-2 sm:order-none"
+          link="markets"
+          title="Go To Markets"
+          handleLeftClick={handleLeftClick}
+          handleRightClick={handleRightClick}
+          rightDisabled={rightDisabled}
+          leftDisabled={leftDisabled}
+        />
+        {/* <div className="flex items-center mb-ztg-30">
+        </div> */}
+        <div className="sm:col-span-2 relative">
           {(scrollDirection === "left" && scrollLeft !== 0) ||
           (scrollDirection === "right" && hasReachedEnd) ? (
-            <div className="bg-gradient-to-r from-white h-[175px] w-[20px] absolute z-ztg-10 -left-[5px]"></div>
+            <div className="bg-gradient-to-r from-white w-[20px] absolute z-ztg-10 -left-[5px]"></div>
           ) : (
-            <div className="bg-gradient-to-r from-transparent to-white h-[175px] w-[20px] absolute z-ztg-10 -right-[5px]"></div>
+            <div className="bg-gradient-to-r from-transparent to-white w-[20px] absolute z-ztg-10 -right-[5px]"></div>
           )}
           <div
             ref={scrollRef}
-            className="flex h-[175px] gap-x-[30px] no-scroll-bar overflow-x-auto"
+            className="flex flex-col md:flex-row gap-4 no-scroll-bar overflow-x-auto"
           >
-            {markets.map((market) => (
-              <MarketCard
-                key={market.marketId}
-                {...market}
-                className="market-card bg-anti-flash-white rounded-ztg-10 min-w-[320px] w-[320px] transition duration-500 ease-in-out"
-              />
-            ))}
+            {markets.map(
+              (market, index) =>
+                index < cardsShown && (
+                  <MarketCard
+                    key={market.marketId}
+                    {...market}
+                    className="market-card bg-anti-flash-white rounded-ztg-10 transition duration-500 ease-in-out"
+                  />
+                ),
+            )}
           </div>
         </div>
       </div>
