+++ components/markets/market-card/index.tsx
@@ -30,7 +30,7 @@ const MarketCardInfoRow = ({
   value?: string;
 }) => {
   return (
-    <div className="mb-1 line-clamp-1">
+    <div className="">
       <span className="text-sky-600">{name}:</span>{" "}
       {value == null ? (
         <Skeleton
@@ -48,11 +48,14 @@ const MarketCardInfoRow = ({
 
 const MarketCardInfo = ({
   rows,
+  question,
 }: {
   rows: { name: string; value: string }[];
+  question: string;
 }) => {
   return (
-    <div className="w-full h-full flex flex-col  justify-between text-ztg-12-120 mt-[10px]">
+    <div className="pl-[15px] w-full h-full flex flex-col justify-center text-ztg-14-165">
+      <h5 className="black font-bold w-full h-fit">{question}</h5>
       {rows.map((r, idx) => (
         <MarketCardInfoRow {...r} key={idx} />
       ))}
@@ -79,15 +82,12 @@ const MarketCard = ({
       name: "Volume",
       value: `${volume ?? 0} ${baseAsset?.toUpperCase() ?? "ZTG"}`,
     },
-    { name: "Status", value: creation },
+    // { name: "Status", value: creation },
   ];
   return (
     <MarketCardContext.Provider value={{ baseAsset }}>
       <div
-        className={
-          "w-full h-full bg-anti-flash-white rounded-[10px] p-[15px] flex flex-col relative " +
-          className
-        }
+        className={`flex flex-col justify-center w-full bg-anti-flash-white rounded-[10px] p-[15px] relative ${className}`}
         data-testid={`marketCard-${marketId}`}
       >
         {showDetailsOverlay && (
@@ -98,22 +98,16 @@ const MarketCard = ({
             onCloseIconClick={() => setShowDetailsOverlay(false)}
           />
         )}
-        {outcomes?.length > 0 && (
+        {/* {outcomes?.length > 0 && (
           <MoreVertical
             className="absolute right-[10px] text-pastel-blue cursor-pointer"
             onClick={() => setShowDetailsOverlay(true)}
           />
-        )}
-        <Link
-          href={`/markets/${marketId}`}
-          className="flex flex-row mb-3 mr-[17px]"
-        >
+        )} */}
+        <Link href={`/markets/${marketId}`} className="flex items-center">
           <MarketImage image={img} alt={question} />
-          <div className="ml-[15px] black font-bold w-full h-fit line-clamp-3 text-ztg-14-150">
-            {question}
-          </div>
+          <MarketCardInfo question={question} rows={infoRows} />
         </Link>
-        <MarketCardInfo rows={infoRows} />
       </div>
     </MarketCardContext.Provider>
   );
