+++ lib/gql/newest-markets.ts
@@ -1,6 +1,6 @@
 import { ZTG } from "@zeitgeistpm/sdk-next";
 import { MarketCreation } from "@zeitgeistpm/sdk/dist/types";
-import { IndexedMarketCardData } from "components/markets/market-card";
+import { IndexedMarketCardData } from "components/markets/market-card/index";
 import Decimal from "decimal.js";
 import { gql, GraphQLClient } from "graphql-request";
 import { MarketOutcomes, MarketOutcome } from "lib/types/markets";
