+++ pages/index-new.tsx
@@ -1,7 +1,8 @@
 import LearnSection from "components/front-page/LearnSection";
 import PopularCategories from "components/front-page/PopularCategories";
-import { IndexedMarketCardData } from "components/markets/market-card";
+import { IndexedMarketCardData } from "components/markets/market-card/index";
 import MarketScroll from "components/markets/MarketScroll";
+import MarketScrollNew from "components/markets/MarketScroll";
 import { motion } from "framer-motion";
 import { GraphQLClient } from "graphql-request";
 import getFeaturedMarkets from "lib/gql/featured-markets";
